+++ bevy_replicon_renet2/examples/tic_tac_toe.rs
@@ -238,14 +238,10 @@ impl TicTacToePlugin {
                 game_state.set(GameState::InGame);
             }
             Cli::Server { port, symbol } => {
-                let server_channels_config = channels.get_server_configs();
-                let client_channels_config = channels.get_client_configs();
-
-                let server = RenetServer::new(ConnectionConfig {
-                    server_channels_config,
-                    client_channels_config,
-                    ..Default::default()
-                });
+                let server = RenetServer::new(ConnectionConfig::new_with_channels(
+                    channels.get_server_configs(),
+                    channels.get_client_configs(),
+                ));
 
                 let current_time = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH)?;
                 let public_addr = SocketAddr::new(Ipv4Addr::LOCALHOST.into(), port);
@@ -264,14 +260,10 @@ impl TicTacToePlugin {
                 commands.spawn(PlayerBundle::server(symbol));
             }
             Cli::Client { port, ip } => {
-                let server_channels_config = channels.get_server_configs();
-                let client_channels_config = channels.get_client_configs();
-
-                let client = RenetClient::new(ConnectionConfig {
-                    server_channels_config,
-                    client_channels_config,
-                    ..Default::default()
-                });
+                let client = RenetClient::new(ConnectionConfig::new_with_channels(
+                    channels.get_server_configs(),
+                    channels.get_client_configs(),
+                ));
 
                 let current_time = SystemTime::now().duration_since(SystemTime::UNIX_EPOCH)?;
                 let client_id = current_time.as_millis() as u64;